Fix arrow rendering
When creating the qwc-services.github.io docu site the generator would leave &rarr; as is and that would render in the browser also as &rarr; (not as an arrow, but as that character sequence).

The qwc-services.github.io page sets `<meta charset="utf-8">`. So instead of using html codes, we can just use the utf-8 character for the arrow.